Fix module registration under browserify
for reference: 850b8979, 113bf9

When bundled with browserify, module was going down the "browser
global" path, not the "commonjs" path.  The reason was that it was
looking for module as "global.module", but browserify supplies it as an
argument, not a global.  History digging shows that "global." was added
to silence jshint warnings, so I adjusted the environment detection not
to use "global.", and adjusted the jshint rules to prevent it from
complaining about those lines.