Fix use of shrinkToFit together with drawOutOfBounds false
When having big words on a small grid, using shrinkToFit: true should,
together with drawOutOfBounds: false, ensure, that the words are all rendered
within the available space. This should be done by shrinking the words
as much as needed to fit the available space.

However, when drawOutOfBounds is set to false, the code before this commit
simply did not render the word at all if it did not fit the grid (anymore).

This commit fixes the behaviour so that when setting shrinkToFit to true and
drawOutOfBounds to false will correctly render the words on the available space.

Co-authored-by: Arne <arne.wohlert@oss.volkswagen.com>